+++ src/stages/memory.h
@@ -10,7 +10,7 @@ public:
     // Runs the memory stage
     static bool Do();
 
-    static uint64_t GetStat();
+    static uint8_t GetStat();
     // Selects memory address
     static uint64_t GetMemAddress(uint8_t icode);
     // Sets read control signal
@@ -23,7 +23,7 @@ public:
 protected:
     static bool PrintErrorMessage(const int error_code);
 
-    static uint64_t stat_;
+    static uint8_t stat_;
 };
 
 }  // namespace stages
